+++ model/mmwave-mac-csched-sap.h
@@ -247,6 +247,12 @@ public:
 
   virtual void CschedUeReleaseReq (const struct CschedUeReleaseReqParameters& params) = 0;
 
+  /**
+   * \brief Sets the number of resource blocks per resource block group
+   * \param numRbPerRbg number of RB per RBG
+   */
+  virtual void SetNumRbPerRbg (uint32_t numRbPerRbg) = 0;
+
 private:
 };