+++ model/mmwave-mac-scheduler-ns3.h
@@ -307,6 +307,8 @@ public:
   DoSchedSetMcs (uint32_t mcs) override;
   virtual void
   DoSchedDlRachInfoReq (const MmWaveMacSchedSapProvider::SchedDlRachInfoReqParameters& params) override;
+  virtual void
+  DoSetNumRbPerRbg (uint32_t numRbPerRbg) override;
 
   // to save some typing
   using HarqVectorIterator = MmWaveMacHarqVector::iterator;
@@ -539,6 +541,14 @@ protected:
   CreateLC (const LogicalChannelConfigListElement_s &config) const;
 
   /**
+   * \brief Private function that is used to get the number of resource
+   * blocks per resource block group and also to check whether this value is
+   * configured.
+   * \return Returns the number of RBs per RBG
+   */
+  int64_t GetNumRbPerRbg () const;
+
+  /**
    * \brief Represent an assignation of bytes to a LCG/LC
    */
   struct Assignation
@@ -712,6 +722,8 @@ private:
   uint8_t m_startMcsDl   {0};   //!< Starting (or fixed) value for DL MCS
   uint8_t m_startMcsUl   {0};   //!< Starting (or fixed) value for UL MCS
   Time    m_cqiTimersThreshold; //!< The time while a CQI is valid
+
+  int32_t m_numRbPerRbg {-1};   //!< number of resource blocks within the channel bandwidth
 
   MmWaveMacSchedulerCQIManagement m_cqiManagement; //!< CQI Management