+++ model/mmwave-phy.cc
@@ -60,6 +60,8 @@ public:
 
   virtual void SetSlotAllocInfo (SlotAllocInfo slotAllocInfo) override;
 
+  virtual void SetNumRbPerRbg (uint32_t numRbPerRbg);
+
   virtual BeamId GetBeamId (uint8_t rnti) const override;
 
   virtual Ptr<const SpectrumModel> GetSpectrumModel () const override;
@@ -98,6 +100,12 @@ void
 MmWaveMemberPhySapProvider::SetSlotAllocInfo (SlotAllocInfo slotAllocInfo)
 {
   m_phy->PushBackSlotAllocInfo (slotAllocInfo);
+}
+
+void
+MmWaveMemberPhySapProvider::SetNumRbPerRbg (uint32_t numRbPerRbg)
+{
+  m_phy->DoSetNumRbPerRbg(numRbPerRbg);
 }
 
 BeamId
@@ -142,14 +150,14 @@ MmWavePhy::FromRBGBitmaskToRBAssignment (const std::vector<uint8_t> rbgBitmask) 
     {
       if (rbgBitmask.at (i) == 1)
         {
-          for (uint32_t k = 0; k < m_phyMacConfig->GetNumRbPerRbg (); ++k)
+          for (uint32_t k = 0; k < GetNumRbPerRbg (); ++k)
             {
-              ret.push_back ((i * m_phyMacConfig->GetNumRbPerRbg ()) + k);
+              ret.push_back ((i * GetNumRbPerRbg ()) + k);
             }
         }
     }
 
-  NS_ASSERT (static_cast<uint32_t> (std::count (rbgBitmask.begin (), rbgBitmask.end (), 1) * m_phyMacConfig->GetNumRbPerRbg ())
+  NS_ASSERT (static_cast<uint32_t> (std::count (rbgBitmask.begin (), rbgBitmask.end (), 1) * GetNumRbPerRbg ())
              == ret.size ());
   return ret;
 }
@@ -266,6 +274,19 @@ MmWavePhy::GetPacketBurst (SfnSf sfn)
       m_packetBurstMap.erase (it);
     }
   return pburst;
+}
+
+void
+MmWavePhy::DoSetNumRbPerRbg (int32_t numRbPerRbg)
+{
+  NS_ABORT_MSG("This function should not be called. GNB PHY should override this function, while this function of UE PHY should not be called.");
+}
+
+int64_t
+MmWavePhy::GetNumRbPerRbg () const
+{
+  NS_ABORT_MSG_IF(m_numRbPerRbg == -1, "Number of RBs per RBG not configured.");
+  return m_numRbPerRbg;
 }
 
 Ptr<SpectrumValue>