+++ model/mmwave-phy.h
@@ -134,6 +134,13 @@ public:
    */
   void EncodeCtrlMsg (const Ptr<MmWaveControlMessage> &msg);
 
+  /**
+   * \brief This function is called by MmWaveMemberPhySapProvider (by MAC) and
+   * it configuress the number of resource blocks per resource block group
+   * \param numRbPerRbg number of resource block per RBG
+   */
+  virtual void DoSetNumRbPerRbg (int32_t numRbPerRbg);
+
 protected:
 
   /**
@@ -151,6 +158,14 @@ protected:
    * is a 1 are from the 4th to the 11th, and that is reflected in the output)
    */
   std::vector<int> FromRBGBitmaskToRBAssignment (const std::vector<uint8_t> rbgBitmask) const;
+
+  /**
+   * \brief Protected function that is used to get the number of resource
+   * blocks per resource block group and also to check whether this value is
+   * configured.
+   * \return Returns the number of RBs per RBG
+   */
+  virtual int64_t GetNumRbPerRbg () const;
 
   Ptr<PacketBurst> GetPacketBurst (SfnSf);
 
@@ -288,6 +303,8 @@ protected:
   Ptr<BeamManager> m_beamManager; //!< TODO
 
   std::list <Ptr<MmWaveControlMessage>> m_ctrlMsgs; //!< CTRL messages to be sent
+
+  int64_t m_numRbPerRbg {-1};   //!< number of resource blocks within the channel bandwidth, this parameter is configured by MAC through phy SAP provider interface
 
 private:
   std::list<SlotAllocInfo> m_slotAllocInfo; //!< slot allocation info list